updated order total to correct data format

+++ Back-end/OrdersAPI/Models/Order.cs
@@ -8,7 +8,7 @@ namespace OrdersAPI.Models
         public int OrderId { get; set; }
         public int ProductId { get; set; }
         public DateTime OrderDate { get; set; }
-        public int OrderTotal { get; set; }
+        public decimal OrderTotal { get; set; }
         public bool OrderCancellation { get; set; }
     }
 }
