+++ Front-End/IrishNFTs.MVC/Models/OrderViewModel.cs
@@ -6,7 +6,7 @@ namespace IrishNFTs.MVC.Models
         public int OrderId { get; set; }
         public int ProductId { get; set; }
         public DateTime OrderDate { get; set; }
-        public int OrderTotal { get; set; }
+        public decimal OrderTotal { get; set; }
         public bool OrderCancellation { get; set; }
